Addressing PR Comments: Created new files to separate code. Resolved naming issues.
