+++ firmware/Core/Src/gps/gps.c
@@ -1,13 +1,5 @@
 #include "gps/gps.h"
-#include "gps/gps_types.h"
-#include "log/log.h"
-#include "telecommands/telecommand_args_helpers.h"
-#include "transforms/arrays.h"
-#include "debug_tools/debug_uart.h"
 
-#include <stdio.h>
-#include <stdlib.h> 
-#include <string.h>
 #include <stdint.h>
 
 #define BUFFER_SIZE 256
@@ -49,897 +41,4 @@ uint32_t GPS_calculate_block_crc32( uint32_t ulCount, uint8_t *ucBuffer ) {
         }
 
     return( crc );
-}
-
-
-/// @brief Assigns a GPS time status based on the provided string.
-/// @param status_str The status string to parse.
-/// @param status Pointer to GPS_reference_time_status_t where the status will be stored.
-/// @return Returns 0 on success, 1 if the status string is unrecognized.
-uint8_t GPS_reference_time_status_str_to_enum(const char *status_str, GPS_reference_time_status_t *status) {
-    if (strcmp(status_str, "UNKNOWN") == 0) {
-        *status = GPS_REF_TIME_UNKNOWN;
-    } else if (strcmp(status_str, "APPROXIMATE") == 0) {
-        *status = GPS_REF_TIME_APPROXIMATE;
-    } else if (strcmp(status_str, "COARSEADJUSTING") == 0) {
-        *status = GPS_REF_TIME_COARSEADJUSTING;
-    } else if (strcmp(status_str, "COARSE") == 0) {
-        *status = GPS_REF_TIME_COARSE;
-    } else if (strcmp(status_str, "COARSESTEERING") == 0) {
-        *status = GPS_REF_TIME_COARSESTEERING;
-    } else if (strcmp(status_str, "FREEWHEELING") == 0) {
-        *status = GPS_REF_TIME_FREEWHEELING;
-    } else if (strcmp(status_str, "FINEADJUSTING") == 0) {
-        *status = GPS_REF_TIME_FINEADJUSTING;
-    } else if (strcmp(status_str, "FINE") == 0) {
-        *status = GPS_REF_TIME_FINE;
-    } else if (strcmp(status_str, "FINEBACKUPSTEERING") == 0) {
-        *status = GPS_REF_TIME_FINEBACKUPSTEERING;
-    } else if (strcmp(status_str, "FINESTEERING") == 0) {
-        *status = GPS_REF_TIME_FINESTEERING;
-    } else if (strcmp(status_str, "SATTIME") == 0) {
-        *status = GPS_REF_TIME_SATTIME;
-    } else {
-        return 1;  // Unrecognized status string
-    }
-    return 0;  // Success
-}
-
-/// @brief Assigns a string value based on the provided GPS time status.
-/// @param status GPS_reference_time_status_t value.
-/// @return Returns the assigned string value for the GPS time status.
-const char* GPS_reference_time_status_enum_to_str(GPS_reference_time_status_t status) {
-    switch (status) {
-        case GPS_REF_TIME_UNKNOWN:
-            return "UNKNOWN";
-        case GPS_REF_TIME_APPROXIMATE:
-            return "APPROXIMATE";
-        case GPS_REF_TIME_COARSEADJUSTING:
-            return "COARSEADJUSTING";
-        case GPS_REF_TIME_COARSE:
-            return "COARSE";
-        case GPS_REF_TIME_COARSESTEERING:
-            return "COARSESTEERING";
-        case GPS_REF_TIME_FREEWHEELING:
-            return "FREEWHEELING";
-        case GPS_REF_TIME_FINEADJUSTING:
-            return "FINEADJUSTING";
-        case GPS_REF_TIME_FINE:
-            return "FINE";
-        case GPS_REF_TIME_FINEBACKUPSTEERING:
-            return "FINEBACKUPSTEERING";
-        case GPS_REF_TIME_FINESTEERING:
-            return "FINESTEERING";
-        case GPS_REF_TIME_SATTIME:
-            return "SATTIME";
-        default:
-            return "UNKNOWN STATUS";  // If status is unrecognized
-    }
-}
-
-
-/// @brief Parse the received GPS header into a struct
-/// @param data_received - The string obtained from the buffer that is to be parsed into the GPS_header_response_t struct
-/// @param result - GPS_header_response_t struct that is returned
-/// @return 0 if successful, > 0 if an error occurred
-uint8_t GPS_header_response_parser(const char *data_received, GPS_header_response_t *result){
-
-    // TODO: What if there are multiple responses in the string?
-
-    // Check if the buffer is empty
-    if (data_received == NULL || data_received[0] == '\0') {
-        // Empty or NULL string, return an error
-        return 1;
-    }
-
-    // Find the start and end of the header, which is # and ; resepectively
-    const char *sync_char = strchr(data_received,'#');
-    const char *delimiter_char = strchr(data_received,';');
-
-    if (!sync_char || !delimiter_char) {
-        // Invalid data: No header in gps response
-        return 2; 
-    }
-
-    // Calculate the length of the header string
-    const int header_length = delimiter_char - sync_char + 1;
-    if (header_length < 0) {
-        //Sync character occurs after delimiter character
-        return 3;
-    }
-
-    char header_buffer[256];
-    if ((size_t)header_length >= sizeof(header_buffer)) {
-        //Header is too large for the buffer
-        return 4;  
-    }
-
-    // Copy header string into a buffer
-    strncpy(header_buffer, sync_char, header_length);
-    header_buffer[header_length] = '\0';  // Null-terminate the substring
-
-    // Parse the data in the header buffer
-    uint8_t parse_result;
-    char token_buffer[128];
-
-    // Log Name
-    parse_result = TCMD_extract_string_arg(header_buffer, 0, token_buffer, sizeof(token_buffer));
-    if (parse_result != 0) {  
-        return parse_result;  
-    }
-    strcpy(result->log_name, token_buffer + 1);
-
-    // Time Status
-    parse_result = TCMD_extract_string_arg(header_buffer, 4, token_buffer, sizeof(token_buffer));
-    if (parse_result != 0) {  
-        return parse_result;  
-    }
-    const uint8_t status_result = GPS_reference_time_status_str_to_enum(token_buffer, &result->time_status);
-    if (status_result != 0) {
-        // Time Status not recognized
-        return status_result;
-    }
-
-    char message_buffer[256];
-    snprintf(
-        message_buffer, sizeof(message_buffer),
-        "{\"log_name\":\"%s\",\"time_status\":\"%s\"}\n",
-        result->log_name,
-        GPS_reference_time_status_enum_to_str(result->time_status)
-    );
-    
-    LOG_message(
-            LOG_SYSTEM_TELECOMMAND, LOG_SEVERITY_NORMAL, LOG_SINK_ALL,
-            "Header Response: %s",
-            message_buffer
-        );
-        
-    return 0;
-}
-
-
-/// @brief Assigns a GPS solution status based on the provided string.
-/// @param status_str The status string to parse.
-/// @param status Pointer to GPS_solution_status_enum_t where the status will be stored.
-/// @return Returns 0 on success, 1 if the status string is unrecognized.
-uint8_t GPS_solution_status_str_to_enum(const char *status_str, GPS_solution_status_enum_t *status) {
-    if (strcmp(status_str, "SOL_COMPUTED") == 0) {
-        *status = GPS_SOL_COMPUTED;
-    } else if (strcmp(status_str, "INSUFFICIENT_OBS") == 0) {
-        *status = GPS_INSUFFICIENT_OBS;
-    } else if (strcmp(status_str, "NO_CONVERGENCE") == 0) {
-        *status = GPS_NO_CONVERGENCE;
-    } else if (strcmp(status_str, "SINGULARITY") == 0) {
-        *status = GPS_SINGULARITY;
-    } else if (strcmp(status_str, "COV_TRACE") == 0) {
-        *status = GPS_COV_TRACE;
-    } else if (strcmp(status_str, "TEST_DIST") == 0) {
-        *status = GPS_TEST_DIST;
-    } else if (strcmp(status_str, "COLD_START") == 0) {
-        *status = GPS_COLD_START;
-    } else if (strcmp(status_str, "V_H_LIMIT") == 0) {
-        *status = GPS_V_H_LIMIT;
-    } else if (strcmp(status_str, "VARIANCE") == 0) {
-        *status = GPS_VARIANCE;
-    } else if (strcmp(status_str, "RESIDUALS") == 0) {
-        *status = GPS_RESIDUALS;
-    } else if (strcmp(status_str, "INTEGRITY_WARNING") == 0) {
-        *status = GPS_INTEGRITY_WARNING;
-    } else {
-        return 1;  // Unrecognized status string
-    }
-    return 0;  // Success
-}
-
-
-/// @brief Assigns a GPS position or velocity type based on the provided string.
-/// @param type_str The type string to parse.
-/// @param type Pointer to GPS_position_velocity_type_enum_t where the type will be stored.
-/// @return Returns 0 on success, 1 if the type string is unrecognized.
-uint8_t GPS_position_type_str_to_enum(const char *type_str, GPS_position_type_enum_t *type) {
-    if (strcmp(type_str, "NONE") == 0) {
-        *type = GPS_TYPE_NONE;
-    } else if (strcmp(type_str, "FIXEDPOS") == 0) {
-        *type = GPS_TYPE_FIXEDPOS;
-    } else if (strcmp(type_str, "FIXEDHEIGHT") == 0) {
-        *type = GPS_TYPE_FIXEDHEIGHT;
-    } else if (strcmp(type_str, "DOPPLER_VELOCITY") == 0) {
-        *type = GPS_TYPE_DOPPLER_VELOCITY;
-    } else if (strcmp(type_str, "SINGLE") == 0) {
-        *type = GPS_TYPE_SINGLE;
-    } else if (strcmp(type_str, "PSDIFF") == 0) {
-        *type = GPS_TYPE_PSDIFF;
-    } else if (strcmp(type_str, "WAAS") == 0) {
-        *type = GPS_TYPE_WAAS;
-    } else if (strcmp(type_str, "PROPAGATED") == 0) {
-        *type = GPS_TYPE_PROPAGATED;
-    } else if (strcmp(type_str, "L1_FLOAT") == 0) {
-        *type = GPS_TYPE_L1_FLOAT;
-    } else if (strcmp(type_str, "NARROW_FLOAT") == 0) {
-        *type = GPS_TYPE_NARROW_FLOAT;
-    } else if (strcmp(type_str, "L1_INT") == 0) {
-        *type = GPS_TYPE_L1_INT;
-    } else if (strcmp(type_str, "WIDE_INT") == 0) {
-        *type = GPS_TYPE_WIDE_INT;
-    } else if (strcmp(type_str, "NARROW_INT") == 0) {
-        *type = GPS_TYPE_NARROW_INT;
-    } else if (strcmp(type_str, "RTK_DIRECT_INS") == 0) {
-        *type = GPS_TYPE_RTK_DIRECT_INS;
-    } else if (strcmp(type_str, "INS_SBAS") == 0) {
-        *type = GPS_TYPE_INS_SBAS;
-    } else if (strcmp(type_str, "INS_PSRSP") == 0) {
-        *type = GPS_TYPE_INS_PSRSP;
-    } else if (strcmp(type_str, "INS_PSRDIFF") == 0) {
-        *type = GPS_TYPE_INS_PSRDIFF;
-    } else if (strcmp(type_str, "INS_RTKFLOAT") == 0) {
-        *type = GPS_TYPE_INS_RTKFLOAT;
-    } else if (strcmp(type_str, "INS_RTKFIXED") == 0) {
-        *type = GPS_TYPE_INS_RTKFIXED;
-    } else if (strcmp(type_str, "PPP_CONVERGING") == 0) {
-        *type = GPS_TYPE_PPP_CONVERGING;
-    } else if (strcmp(type_str, "PPP") == 0) {
-        *type = GPS_TYPE_PPP;
-    } else if (strcmp(type_str, "OPERATIONAL") == 0) {
-        *type = GPS_TYPE_OPERATIONAL;
-    } else if (strcmp(type_str, "WARNING") == 0) {
-        *type = GPS_TYPE_WARNING;
-    } else if (strcmp(type_str, "OUT_OF_BOUNDS") == 0) {
-        *type = GPS_TYPE_OUT_OF_BOUNDS;
-    } else if (strcmp(type_str, "INS_PPP_CONVERGING") == 0) {
-        *type = GPS_TYPE_INS_PPP_CONVERGING;
-    } else if (strcmp(type_str, "INS_PPP") == 0) {
-        *type = GPS_TYPE_INS_PPP;
-    } else if (strcmp(type_str, "PPP_BASIC_CONVERGING") == 0) {
-        *type = GPS_TYPE_PPP_BASIC_CONVERGING;
-    } else if (strcmp(type_str, "PPP_BASIC") == 0) {
-        *type = GPS_TYPE_PPP_BASIC;
-    } else if (strcmp(type_str, "INS_PPP_BASIC_CONVERGING") == 0) {
-        *type = GPS_TYPE_INS_PPP_BASIC_CONVERGING;
-    } else if (strcmp(type_str, "INS_PPP_BASIC") == 0) {
-        *type = GPS_TYPE_INS_PPP_BASIC;
-    } else {
-        return 1;  // Unrecognized type string
-    }
-    return 0;  // Success
-}
-
-/// @brief Assigns a string value based on the provided GPS solution status.
-/// @param status GPS_solution_status_enum_t value.
-/// @return Returns the assigned string value for the GPS_solution_status_enum_t.
-const char* GPS_position_type_enum_to_str(GPS_solution_status_enum_t status) {
-    switch (status) {
-        case GPS_SOL_COMPUTED:
-            return "SOL_COMPUTED";
-        case GPS_INSUFFICIENT_OBS:
-            return "INSUFFICIENT_OBS";
-        case GPS_NO_CONVERGENCE:
-            return "NO_CONVERGENCE";
-        case GPS_SINGULARITY:
-            return "SINGULARITY";
-        case GPS_COV_TRACE:
-            return "COV_TRACE";
-        case GPS_TEST_DIST:
-            return "TEST_DIST";
-        case GPS_COLD_START:
-            return "COLD_START";
-        case GPS_V_H_LIMIT:
-            return "V_H_LIMIT";
-        case GPS_VARIANCE:
-            return "VARIANCE";
-        case GPS_RESIDUALS:
-            return "RESIDUALS";
-        case GPS_INTEGRITY_WARNING:
-            return "INTEGRITY_WARNING";
-        default:
-            return "UNKNOWN STATUS";  // For unrecognized enum values
-    }
-}
-
-/// @brief Assigns a string value based on the provided GPS position type.
-/// @param type GPS_position_type_enum_t value.
-/// @return Returns the assigned string value for the GPS_position_type_enum_t.
-const char* GPS_position_type_enum_to_string(GPS_position_type_enum_t type) {
-    switch (type) {
-        case GPS_TYPE_NONE:
-            return "NONE";
-        case GPS_TYPE_FIXEDPOS:
-            return "FIXEDPOS";
-        case GPS_TYPE_FIXEDHEIGHT:
-            return "FIXEDHEIGHT";
-        case GPS_TYPE_DOPPLER_VELOCITY:
-            return "DOPPLER_VELOCITY";
-        case GPS_TYPE_SINGLE:
-            return "SINGLE";
-        case GPS_TYPE_PSDIFF:
-            return "PSDIFF";
-        case GPS_TYPE_WAAS:
-            return "WAAS";
-        case GPS_TYPE_PROPAGATED:
-            return "PROPAGATED";
-        case GPS_TYPE_L1_FLOAT:
-            return "L1_FLOAT";
-        case GPS_TYPE_NARROW_FLOAT:
-            return "NARROW_FLOAT";
-        case GPS_TYPE_L1_INT:
-            return "L1_INT";
-        case GPS_TYPE_WIDE_INT:
-            return "WIDE_INT";
-        case GPS_TYPE_NARROW_INT:
-            return "NARROW_INT";
-        case GPS_TYPE_RTK_DIRECT_INS:
-            return "RTK_DIRECT_INS";
-        case GPS_TYPE_INS_SBAS:
-            return "INS_SBAS";
-        case GPS_TYPE_INS_PSRSP:
-            return "INS_PSRSP";
-        case GPS_TYPE_INS_PSRDIFF:
-            return "INS_PSRDIFF";
-        case GPS_TYPE_INS_RTKFLOAT:
-            return "INS_RTKFLOAT";
-        case GPS_TYPE_INS_RTKFIXED:
-            return "INS_RTKFIXED";
-        case GPS_TYPE_PPP_CONVERGING:
-            return "PPP_CONVERGING";
-        case GPS_TYPE_PPP:
-            return "PPP";
-        case GPS_TYPE_OPERATIONAL:
-            return "OPERATIONAL";
-        case GPS_TYPE_WARNING:
-            return "WARNING";
-        case GPS_TYPE_OUT_OF_BOUNDS:
-            return "OUT_OF_BOUNDS";
-        case GPS_TYPE_INS_PPP_CONVERGING:
-            return "INS_PPP_CONVERGING";
-        case GPS_TYPE_INS_PPP:
-            return "INS_PPP";
-        case GPS_TYPE_PPP_BASIC_CONVERGING:
-            return "PPP_BASIC_CONVERGING";
-        case GPS_TYPE_PPP_BASIC:
-            return "PPP_BASIC";
-        case GPS_TYPE_INS_PPP_BASIC_CONVERGING:
-            return "INS_PPP_BASIC_CONVERGING";
-        case GPS_TYPE_INS_PPP_BASIC:
-            return "INS_PPP_BASIC";
-        default:
-            return "UNKNOWN TYPE";  // If type is unrecognized
-    }
-}
-
-
-/// @brief Parse Received Data
-/// @param data_received - Number of bytes in the data block
-/// @return 0 if successful, >0 if an error occurred
-uint8_t GPS_bestxyza_data_parser(const char* data_received, GPS_bestxyza_response_t *result) {
-
-    // Check if the buffer is empty
-    if (data_received[0] == '\0') {
-        // Empty or NULL string, return an error
-        LOG_message(
-            LOG_SYSTEM_TELECOMMAND, LOG_SEVERITY_NORMAL, LOG_SINK_ALL,
-            "Error 1: Empty buffer"
-        );
-        return 1;
-    }
-
-    GPS_header_response_t bestxyza_header;
-    const uint8_t header_parse_result = GPS_header_response_parser(data_received,&bestxyza_header);
-
-    if(header_parse_result != 0){
-        // Error in parsing header section
-        LOG_message(
-            LOG_SYSTEM_TELECOMMAND, LOG_SEVERITY_NORMAL, LOG_SINK_ALL,
-            "Error 2: Error parising the gps header"
-        );
-        return 2;
-    }
-
-    if(strcmp(bestxyza_header.log_name, "BESTXYZA") != 0){
-        // Incorrect log function
-        LOG_message(
-            LOG_SYSTEM_TELECOMMAND, LOG_SEVERITY_NORMAL, LOG_SINK_ALL,
-            "Error 3: Incorrect log function ie not BESTXYZA"
-        );
-        return 3;
-    }
-
-    // TODO: What if there is multiple commands in the string?
-    const char *header_delimiter_char = strchr(data_received,';');
-    const char* bestxyza_data_start = header_delimiter_char + 1;
-    const char* asterisk = strchr(bestxyza_data_start, '*');  
-
-    if(strcmp(bestxyza_data_start, "\0") == 0){
-        // No data after the gps header within the response
-        LOG_message(
-            LOG_SYSTEM_TELECOMMAND, LOG_SEVERITY_NORMAL, LOG_SINK_ALL,
-            "Error 4: Missing Data after the header"
-        );
-        return 4;
-    }  
-
-    if(!asterisk){
-        // No terminator at the end of the bestxyza data, ie no CRC present
-        LOG_message(
-            LOG_SYSTEM_TELECOMMAND, LOG_SEVERITY_NORMAL, LOG_SINK_ALL,
-            "Error 5: Missing Astericks in response"
-        );
-        return 5;
-    }
-
-    const int bestxyza_data_length = asterisk - bestxyza_data_start + 1;
-    char bestxyza_data_buffer[512];
-    if ((size_t)bestxyza_data_length >= sizeof(bestxyza_data_buffer)) {
-        //Buffer Overflow Error
-        LOG_message(
-            LOG_SYSTEM_TELECOMMAND, LOG_SEVERITY_NORMAL, LOG_SINK_ALL,
-            "Error 6: Buffer overflow"
-        );
-        return 6;  
-    }
-
-    strncpy(bestxyza_data_buffer, bestxyza_data_start, bestxyza_data_length);
-    bestxyza_data_buffer[bestxyza_data_length] = '\0';
-
-    // Parse the data in the bestxyza data buffer
-    uint8_t parse_result;
-    char token_buffer[128];
-    char *end_ptr;
-    
-    // Position Solution Status
-    parse_result = TCMD_extract_string_arg(bestxyza_data_buffer, 0, token_buffer, sizeof(token_buffer));
-    if (parse_result != 0) {  
-        return parse_result;  
-    }
-    uint8_t status_result = GPS_solution_status_str_to_enum(token_buffer, &result->position_solution_status);
-    if(status_result != 0){
-        // Invalid string passed
-        return status_result;
-    }
-
-    // Position Type
-    parse_result = TCMD_extract_string_arg(bestxyza_data_buffer, 1, token_buffer, sizeof(token_buffer));
-    if (parse_result != 0) {  
-        return parse_result;  
-    }
-    status_result = GPS_position_type_str_to_enum(token_buffer, &result->position_type);
-    if(status_result != 0){
-        // Invalid string passed
-        return status_result;
-    }
-
-    // Position Coordinates 
-    parse_result = TCMD_extract_string_arg(bestxyza_data_buffer, 2, token_buffer, sizeof(token_buffer));
-    if (parse_result != 0) {  
-        return parse_result;  
-    }
-
-    double value = strtod(token_buffer, &end_ptr);
-    if(strcmp(end_ptr, token_buffer) == 0){
-        // Improper string to double conversion
-        return 7;
-    }
-    double conv_result = value * 1000;
-
-    //TODO: Determine what to do when we exceed data type value
-    if(conv_result > INT64_MAX){
-        conv_result = INT64_MAX;
-    }
-    else if(conv_result < INT64_MIN){
-        conv_result = INT64_MIN;
-    }
-    result->position_x_mm = (int64_t) conv_result;
-
-    parse_result = TCMD_extract_string_arg(bestxyza_data_buffer, 3, token_buffer, sizeof(token_buffer));
-    if (parse_result != 0) {  
-        return parse_result;  
-    }
-    value = strtod(token_buffer, &end_ptr);
-    if(strcmp(end_ptr, token_buffer) == 0){
-        // Improper string to double conversion
-        return 7;
-    }
-    conv_result = value * 1000;
-
-    //TODO: Determine what to do when we exceed data type value
-    if(conv_result > INT64_MAX){
-        conv_result = INT64_MAX;
-    }
-    else if(conv_result < INT64_MIN){
-        conv_result = INT64_MIN;
-    }
-    result->position_y_mm = (int64_t) conv_result;
-
-    parse_result = TCMD_extract_string_arg(bestxyza_data_buffer, 4, token_buffer, sizeof(token_buffer));
-    if (parse_result != 0) {  
-        return parse_result;  
-    }
-    value = strtod(token_buffer, &end_ptr);
-    if(strcmp(end_ptr, token_buffer) == 0){
-        // Improper string to double conversion
-        return 7;
-    }
-    conv_result = value * 1000;
-
-    //TODO: Determine what to do when we exceed data type value
-    if(conv_result > INT64_MAX){
-        conv_result = INT64_MAX;
-    }
-    else if(conv_result < INT64_MIN){
-        conv_result = INT64_MIN;
-    }
-    result->position_z_mm = (int64_t) conv_result;
-
-    // Position Coordinates Standard Deviation
-    parse_result = TCMD_extract_string_arg(bestxyza_data_buffer, 5, token_buffer, sizeof(token_buffer));
-    if (parse_result != 0) {  
-        return parse_result;  
-    }
-    value = strtod(token_buffer, &end_ptr);
-    if(strcmp(end_ptr, token_buffer) == 0){
-        // Improper string to double conversion
-        return 7;
-    }
-    conv_result = value * 1000;
-
-    //TODO: Determine what to do when we exceed data type value
-    if(conv_result > INT32_MAX){
-        conv_result = INT32_MAX;
-    }
-    else if(conv_result < INT32_MIN){
-        conv_result = INT32_MIN;
-    }
-    result->position_x_std_mm = (int32_t) conv_result;
-
-    parse_result = TCMD_extract_string_arg(bestxyza_data_buffer, 6, token_buffer, sizeof(token_buffer));
-    if (parse_result != 0) {  
-        return parse_result;  
-    }
-    value = strtod(token_buffer, &end_ptr);
-    if(strcmp(end_ptr, token_buffer) == 0){
-        // Improper string to double conversion
-        return 7;
-    }
-    conv_result = value * 1000;
-
-    //TODO: Determine what to do when we exceed data type value
-    if(conv_result > INT32_MAX){
-        conv_result = INT32_MAX;
-    }
-    else if(conv_result < INT32_MIN){
-        conv_result = INT32_MIN;
-    }
-    result->position_y_std_mm = (int32_t) conv_result;
-
-    parse_result = TCMD_extract_string_arg(bestxyza_data_buffer, 7, token_buffer, sizeof(token_buffer));
-    if (parse_result != 0) {  
-        return parse_result;  
-    }
-    value = strtod(token_buffer, &end_ptr);
-    if(strcmp(end_ptr, token_buffer) == 0){
-        // Improper string to double conversion
-        return 7;
-    }
-    conv_result = value * 1000;
-
-    //TODO: Determine what to do when we exceed data type value
-    if(conv_result > INT32_MAX){
-        conv_result = INT32_MAX;
-    }
-    else if(conv_result < INT32_MIN){
-        conv_result = INT32_MIN;
-    }
-    result->position_z_std_mm = (int32_t) conv_result;
-
-    // Differential Age
-    parse_result = TCMD_extract_string_arg(bestxyza_data_buffer, 18, token_buffer, sizeof(token_buffer));
-    if (parse_result != 0) {  
-        return parse_result;  
-    }
-    value = strtod(token_buffer, &end_ptr);
-    if(strcmp(end_ptr, token_buffer) == 0){
-        // Improper string to double conversion
-        return 7;
-    }
-    conv_result = value * 1000;
-
-    //TODO: Determine what to do when we exceed data type value
-    if(conv_result > INT64_MAX){
-        conv_result = INT64_MAX;
-    }
-    else if(conv_result < INT64_MIN){
-        conv_result = INT64_MIN;
-    }
-    result->differential_age_ms = (int64_t) conv_result;
-
-    // Solution Age
-    parse_result = TCMD_extract_string_arg(bestxyza_data_buffer, 19, token_buffer, sizeof(token_buffer));
-    if (parse_result != 0) {  
-        return parse_result;  
-    }
-    value = strtod(token_buffer, &end_ptr);
-    if(strcmp(end_ptr, token_buffer) == 0){
-        // Improper string to double conversion
-        return 7;
-    }
-    conv_result = value * 1000;
-
-    //TODO: Determine what to do when we exceed data type value
-    if(conv_result > INT64_MAX){
-        conv_result = INT64_MAX;
-    }
-    else if(conv_result < INT64_MIN){
-        conv_result = INT64_MIN;
-    }
-    result->solution_age_ms = (int64_t) conv_result;
-
-    char crc[9];
-    strncpy(crc, asterisk + 1, 8);
-    crc[sizeof(crc)-1] = '\0';
-    result->crc = strtoul(crc,&end_ptr, 16);
-
-    // TODO: Add a check for the CRC
-
-
-    char sol_age[32];
-    GEN_uint64_to_str(result->solution_age_ms, sol_age);
-    char diff_age[32];
-    GEN_uint64_to_str(result->differential_age_ms, diff_age);
-    char pos_x[32];
-    GEN_int64_to_str(result->position_x_mm,pos_x);
-    char pos_y[32];
-    GEN_int64_to_str(result->position_y_mm,pos_y);
-    char pos_z[32];
-    GEN_int64_to_str(result->position_z_mm,pos_z);
-
-    char message_buffer[2048];
-    snprintf(
-        message_buffer, sizeof(message_buffer),
-        "{\"Position Solution Status\":\"%s\",\"Position Type\":\"%s\",\"Position x in mm\":\"%s\",\"Position y in mm\":\"%s\",\"Position z in mm\":\"%s\","
-        "\"Position x std in mm\":\"%lu\",\"Position y std in mm\":\"%lu\",\"Position z std in mm\":\"%lu\",\"Solution Age in ms\":\"%s\",\"Differential age in ms\":\"%s\",\"CRC\":\"%ld\"}\n",
-        GPS_position_type_enum_to_str(result->position_solution_status),
-        GPS_position_type_enum_to_string(result->position_type),
-        pos_x,
-        pos_y,
-        pos_z,
-        result->position_x_std_mm,
-        result->position_y_std_mm,
-        result->position_z_std_mm,
-        sol_age,
-        diff_age,
-        result->crc
-    );
-
-    LOG_message(
-            LOG_SYSTEM_TELECOMMAND, LOG_SEVERITY_NORMAL, LOG_SINK_ALL,
-            "Bestxyza Response: %s",
-            message_buffer
-        );
-
-    return 0;
-}
-
-/// @brief Assigns a GPS Clock Model status based on the provided string.
-/// @param status_str The status string to parse.
-/// @param status Pointer to GPS_clock_model_status_enum_t where the status will be stored.
-/// @return Returns 0 on success, 1 if the status string is unrecognized.
-uint8_t GPS_clock_model_status_str_to_enum(const char *status_str, GPS_clock_model_status_enum_t *status) {
-    if (strcmp(status_str, "VALID") == 0) {
-        *status = GPS_CLOCK_VALID;
-    } else if (strcmp(status_str, "CONVERGING") == 0) {
-        *status = GPS_CLOCK_CONVERGING;
-    } else if (strcmp(status_str, "ITERATING") == 0) {
-        *status = GPS_CLOCK_ITERATING;
-    } else if (strcmp(status_str, "INVALID") == 0) {
-        *status = GPS_CLOCK_INVALID;
-    } else {
-        return 1;  // Unrecognized status string
-    }
-    return 0;  // Success
-}
-
-/// @brief Assigns a GPS UTC status based on the provided string.
-/// @param status_str The status string to parse.
-/// @param status Pointer to GPS_utc_status_enum_t where the status will be stored.
-/// @return Returns 0 on success, 1 if the status string is unrecognized.
-uint8_t GPS_utc_status_str_to_enum(const char *status_str, GPS_utc_status_enum_t *status) {
-    if (strcmp(status_str, "INVALID") == 0) {
-        *status = GPS_UTC_INVALID;
-    } else if (strcmp(status_str, "VALID") == 0) {
-        *status = GPS_UTC_VALID;
-    } else if (strcmp(status_str, "WARNING") == 0) {
-        *status = GPS_UTC_WARNING;
-    } else {
-        return 1;  // Unrecognized status string
-    }
-    return 0;  // Success
-}
-
-/// @brief Assigns a string value based on the provided GPS clock model status.
-/// @param status GPS_clock_model_status_enum_t value.
-/// @return Returns the assigned string value for the GPS_clock_model_status_enum_t.
-const char* GPS_clock_model_status_enum_to_string(GPS_clock_model_status_enum_t status) {
-    switch (status) {
-        case GPS_CLOCK_VALID:
-            return "VALID";
-        case GPS_CLOCK_CONVERGING:
-            return "CONVERGING";
-        case GPS_CLOCK_ITERATING:
-            return "ITERATING";
-        case GPS_CLOCK_INVALID:
-            return "INVALID";
-        default:
-            return "UNKNOWN STATUS";  // For unrecognized enum values
-    }
-}
-
-/// @brief Assigns a string value based on the provided GPS UTC status.
-/// @param status GPS_utc_status_enum_t value.
-/// @return Returns the assigned string value for the GPS_utc_status_enum_t.
-const char* GPS_utc_status_enum_to_string(GPS_utc_status_enum_t status) {
-    switch (status) {
-        case GPS_UTC_INVALID:
-            return "INVALID";
-        case GPS_UTC_VALID:
-            return "VALID";
-        case GPS_UTC_WARNING:
-            return "WARNING";
-        default:
-            return "UNKNOWN STATUS";  // For unrecognized enum values
-    }
-}
-
-
-/// @brief Parse Received Data
-/// @param data_received - Number of bytes in the data block
-/// @return 0 if successful, >0 if an error occurred
-uint8_t GPS_timea_data_parser(const char* data_received, GPS_timea_response_t *result) {
-
-    // Check if the buffer is empty
-    if (data_received[0] == '\0') {
-        // Empty or NULL string, return an error
-        LOG_message(
-            LOG_SYSTEM_TELECOMMAND, LOG_SEVERITY_NORMAL, LOG_SINK_ALL,
-            "Error 1: Empty buffer"
-        );
-        return 1;
-    }
-
-    GPS_header_response_t timea_header;
-    const uint8_t header_parse_result = GPS_header_response_parser(data_received,&timea_header);
-
-    if(header_parse_result != 0){
-        // Error in parsing header section
-        LOG_message(
-            LOG_SYSTEM_TELECOMMAND, LOG_SEVERITY_NORMAL, LOG_SINK_ALL,
-            "Error 2: Header Parsing Error"
-        );
-        return 2;
-    }
-
-    if(strcmp(timea_header.log_name, "TIMEA") != 0){
-        // Incorrect log function
-        LOG_message(
-            LOG_SYSTEM_TELECOMMAND, LOG_SEVERITY_NORMAL, LOG_SINK_ALL,
-            "Error 3: Invalid log function ie not TIMEA"
-        );
-        return 3;
-    }
-
-    // TODO: What if there is multiple commands in the string?
-    const char *header_delimiter_char = strchr(data_received,';');
-    const char* timea_data_start = header_delimiter_char + 1;
-    const char* asterisk = strchr(timea_data_start, '*');
-
-    const int timea_data_length = asterisk - timea_data_start + 1;
-
-    if(strcmp(timea_data_start, "\0") == 0){
-        // No data after the gps header within the response
-        LOG_message(
-            LOG_SYSTEM_TELECOMMAND, LOG_SEVERITY_NORMAL, LOG_SINK_ALL,
-            "Error 4: Missing Data after the header"
-        );
-        return 4;
-    }
-
-    if(!asterisk){
-        // No terminator at the end of the bestxyza data, ie no CRC present
-        LOG_message(
-            LOG_SYSTEM_TELECOMMAND, LOG_SEVERITY_NORMAL, LOG_SINK_ALL,
-            "Error 5: Missing Astericks in response"
-        );
-        return 5;
-    }
-
-    // TODO: Change this to be more conservative of storage space
-    char timea_data_buffer[512];
-    if ((size_t)timea_data_length >= sizeof(timea_data_buffer)) {
-        //Header is too large for the buffer
-        LOG_message(
-            LOG_SYSTEM_TELECOMMAND, LOG_SEVERITY_NORMAL, LOG_SINK_ALL,
-            "Error 6: Buffer Oveflow"
-        );
-        return 6;  
-    }
-
-    strncpy(timea_data_buffer, timea_data_start, timea_data_length);
-    timea_data_buffer[timea_data_length] = '\0';
-
-    // Parse the data in the bestxyza data buffer
-    uint8_t parse_result;
-    char token_buffer[128];
-    char *end_ptr;
-    
-    // Clock Model Status
-    parse_result = TCMD_extract_string_arg(timea_data_buffer, 0, token_buffer, sizeof(token_buffer));
-    if (parse_result != 0) {  
-        return parse_result;  
-    }
-    uint8_t status_result = GPS_clock_model_status_str_to_enum(token_buffer, &result->clock_status);
-    if(status_result != 0){
-        // Invalid string passed
-        return status_result;
-    }
-
-    // UTC Offset
-    parse_result = TCMD_extract_string_arg(timea_data_buffer, 3, token_buffer, sizeof(token_buffer));
-    if (parse_result != 0) {  
-        return parse_result;  
-    }
-
-    double value = strtod(token_buffer, &end_ptr);
-    if(strcmp(end_ptr,token_buffer)==0){
-        return 7;
-    }
-    
-    // TODO: Determine what to do when we exceed data limit ie fall out of bounds of valid int64_t values
-    // May have to change to int32_t for optimization.
-    if(value > INT64_MAX){
-        value = INT64_MAX;
-    } else if (value < INT64_MIN){
-        value = INT64_MIN;
-    }
-    result->utc_offset = (int64_t) value;
-
-
-    // UTC Status
-    parse_result = TCMD_extract_string_arg(timea_data_buffer, 10, token_buffer, sizeof(token_buffer));
-    if (parse_result != 0) {  
-        return parse_result;  
-    }
-
-    char *asterisk_pos = strchr(token_buffer, '*');
-    if (asterisk_pos) {
-        *asterisk_pos = '\0';  
-    }
-    status_result = GPS_utc_status_str_to_enum(token_buffer, &result->utc_status);
-    if(status_result != 0){
-        // Invalid string passed
-        return status_result;
-    }
-
-    // TODO: Look into using TCMD_extract_hex_array
-    // TODO: Add a check for the CRC
-    char crc[9];
-    strncpy(crc, asterisk + 1, 8);
-    crc[sizeof(crc)-1] = '\0';
-    result->crc = strtoul(crc,&end_ptr, 16);
-
-    char utc_offset[32];
-    GEN_int64_to_str(result->utc_offset,utc_offset);
-    char message_buffer[1024];
-    snprintf(
-        message_buffer, sizeof(message_buffer),
-        "{\"Clock Status\":\"%s\",\"UTC Status\":%s,\"UTC Offset\":\"%s\",\"CRC\":%lx}\n",
-        GPS_clock_model_status_enum_to_string(result->clock_status),
-        utc_offset,
-        GPS_utc_status_enum_to_string(result->utc_status),
-        result->crc
-    );
-    
-    LOG_message(
-            LOG_SYSTEM_TELECOMMAND, LOG_SEVERITY_NORMAL, LOG_SINK_ALL,
-            "Timea Response: %s",
-            message_buffer
-        );
-
-    return 0;
 }
