+++ firmware/Core/Src/rtos_tasks/rtos_tasks.c
@@ -12,6 +12,7 @@
 #include "log/log.h"
 #include "gps/gps_types.h"
 #include "gps/gps.h"
+#include "gps/gps_ascii_parsers.h"
 
 #include "cmsis_os.h"
 
@@ -198,7 +199,7 @@ void TASK_receive_gps_info(void *argument) {
 
 		// Parsing the gps data
 		GPS_header_response_t gps_header;
-		const u_int8_t gps_header_result = GPS_header_response_parser(latest_gps_response,&gps_header);
+		const uint8_t gps_header_result = GPS_header_response_parser(latest_gps_response,&gps_header);
 
 		// TODO: Figure out what to do after this
 		// Parse may have failed due to incomplete data
